feat(render): rework mermaid node styling and legend (ADR 0038)
Recolors `removed` to a red/dashed style and moves `fan-in` off red
onto a violet stroke plus an `(in:N)` label suffix, so the two classes
never collide on the same color. Replaces the prose legend line
composed by compose_and_post_comment.sh with a self-describing
`Legend` subgraph rendered by render_mermaid/render_mermaid_file_level
itself, styled with the real classDefs so it can't drift out of sync
with the actual node styles.

Splits mermaid.rs's test module into rinkaku-core/src/render/
mermaid_tests/ (by topic: empty/legend, class styling, escaping,
node-budget fallback, removed symbols) to keep the production file
under the file-size warning threshold after the legend/label changes
grew every test's expected output.
